feat: add participants route

+++ src/Journey.Application/UseCases/Trips/Delete/DeleteTripByIdUseCase.cs
@@ -11,11 +11,21 @@ public class DeleteTripByIdUseCase
     {
         var dbContext = new JourneyDbContext();
 
-        var trip = dbContext.Trips.Include(t => t.Activities).FirstOrDefault(t => t.Id == id);
+        var trip = dbContext.Trips.Include(t => t.Activities).Include(t=> t.Participants).FirstOrDefault(t => t.Id == id);
+        var activities = dbContext.Activities.FirstOrDefault(a => a.TripId == id);
+        var participants = dbContext.Participants.FirstOrDefault(p => p.TripId == id);
 
         if (trip is null)
         {
             throw new NotFoundException(ResourceErrorMessages.TRIP_NOT_FOUND);
+        }
+        if (activities is null)
+        {
+            throw new NotFoundException(ResourceErrorMessages.ACTIVITY_NOT_FOUND);
+        }
+        if (participants is null)
+        {
+            throw new NotFoundException(ResourceErrorMessages.PARTICIPANT_NOT_FOUND);
         }
 
         dbContext.Trips.Remove(trip);
